Split pyCA into separate processes
This patch refactors pyCA to separate functionalities and have them
running in separate processes:

- The scheduler will communicate with Opencast to get the scheduled
  events and keep an up-to-date list in the database.
- The capturer will start captures according to the schedule and then
  upload the events.

The long-term goal would be to separate the ingest functionality as well
to enable re-uploads of events as well as a proper implementation of the
user interface.
