+++ pyca/__init__.py
@@ -1,11 +1,18 @@
 #!/usr/bin/env python
 # -*- coding: utf-8 -*-
 '''
-	python-matterhorn-ca
-	~~~~~~~~~~~~~~~~~~~~
+    python-capture-agent
+    ~~~~~~~~~~~~~~~~~~~~
 
-	:copyright: 2014-2015, Lars Kiesow <lkiesow@uos.de>
-	:license: LGPL – see license.lgpl for more details.
+    :copyright: 2014-2015, Lars Kiesow <lkiesow@uos.de>
+    :license: LGPL – see license.lgpl for more details.
 '''
 
+import logging
 
+
+# Set up logging
+logging.basicConfig(level=logging.INFO,
+                    format='%(asctime)s %(levelname)-8s ' +
+                    '[%(filename)s:%(lineno)s:%(funcName)s()] %(message)s',
+                    datefmt='%Y-%m-%d %H:%M:%S')
